Fix poster getting stuck for deleted anchor posts
This commit fixes the issue that arises when the post that is
used as the before anchor is deleted. When this happens, the API
no longer returns any results when it is provided with its fullname.

The new solution to is to instead find new posts by their created_utc
timestamp. While this requires us to receive a bit more data than we
really need, I could not find a better solution for the time being.